Removing metronome calls from interim bars
Now only called once, in the first bar of a passage. 
This fixes the "hiccup" problem where the new calls to the metronome object override the previous one (not actually - they were being stopped at the end of the previous bar -  but in practice that's what it sounds like). Since it interrupted the rhythm often sounded a bit shit.